240805 20:33 - 소통하기 글 삭제 ing

+++ src/main/java/me/jhchoi/ontrack/mapper/TaskMapper.java
@@ -26,7 +26,7 @@ public interface TaskMapper {
     int delAssignee(TaskAssignment taskAssignment);
 
     // 할 일: 파일 첨부
-    Long attachFile(List<TaskFile> taskFile);
+    int attachFile(List<TaskFile> taskFile);
 
     // 할 일: (작성자에 의한) 파일 삭제
     int delFile(Long fileId);
@@ -45,7 +45,7 @@ public interface TaskMapper {
     
 
     // 할 일 상세: 소통하기 글 등록
-    Long addComment(TaskComment taskComment);
+    int addComment(TaskComment taskComment);
 
     // 할 일 상세: 소통하기 글 수정
     Integer editTaskComment(TaskComment editComment);
